bootstrap: on upgrade check for file before deleting it

+++ bootstrap.sh
@@ -368,11 +368,21 @@ doConfig() {
     #
     # This is a destructive process meaning we can't roll back.
     #
-    rm ${HOME}/.git-prompt.sh
-    rm ${HOME}/.git-completion.bash
-    rm ${HOME}/.tigrc.vim
-    rm ${HOME}/bin/git-quick-stats
-    rm ${HOME}/bin/diff-so-fancy
+    if [ -f "${HOME}/.git-prompt.sh" ]; then
+      rm ${HOME}/.git-prompt.sh
+    fi
+    if [ -f "${HOME}/.git-completion.bash" ]; then
+      rm ${HOME}/.git-completion.bash
+    fi
+    if [ -f "${HOME}/.tigrc.vim" ]; then
+      rm ${HOME}/.tigrc.vim
+    fi
+    if [ -f "${HOME}/bin/git-quick-stats" ]; then
+      rm ${HOME}/bin/git-quick-stats
+    fi
+    if [ -f "${HOME}/bin/diff-so-fancy" ]; then
+      rm ${HOME}/bin/diff-so-fancy
+    fi
   fi
 
   if [ ! -f ${HOME}/.git-prompt.sh ]; then
